refactored Game class
